build: extend "make check" with duplicate verification.
Introduce "check_duplicates.sh" which is a duplicate checker. Jdupes is
better and its .deb is only 45KB on Ubuntu Kinetic and could be used
instead but a) its output needs to be parsed (jdupes versions from a few
months ago have added --json), b) I wasn't sure if a new dependency
would surprise some people.

This is only a duplicate detector. When duplicates are found, the
script advises to use jdupes to de-duplicate (or to do it by hand).